+++ sistema/src/main/java/com/hatunvet/sistema/repository/VentaRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -13,16 +14,17 @@ public interface VentaRepository extends JpaRepository<Venta, String> {
 
     List<Venta> findAllByOrderByFechaEmisionDesc();
 
+    // Sincronizado para devolver BigDecimal (Punto 4)
     @Query("SELECT COALESCE(SUM(v.total), 0) FROM Venta v WHERE DATE(v.fechaEmision) = CURRENT_DATE AND v.estado != 'ANULADO'")
-    Double sumVentasHoy();
+    BigDecimal sumVentasHoy();
 
     @Query("SELECT COALESCE(SUM(v.total), 0) FROM Venta v WHERE MONTH(v.fechaEmision) = MONTH(CURRENT_DATE) AND YEAR(v.fechaEmision) = YEAR(CURRENT_DATE) AND v.estado != 'ANULADO'")
-    Double sumVentasMes();
+    BigDecimal sumVentasMes();
 
+    // El COUNT en JPQL siempre devuelve un Long
     @Query("SELECT COUNT(v) FROM Venta v WHERE MONTH(v.fechaEmision) = MONTH(CURRENT_DATE) AND YEAR(v.fechaEmision) = YEAR(CURRENT_DATE) AND v.estado != 'ANULADO'")
-    Integer countVentasMes();
+    Long countVentasMes();
 
-    // ¡Aquí está la magia! Le decimos que busque por "clienteDocumento"
     @Query("SELECT v FROM Venta v WHERE v.clienteDocumento = :numDoc ORDER BY v.fechaEmision DESC")
     List<Venta> findByNumDocOrderByFechaEmisionDesc(@Param("numDoc") String numDoc);
 }
